requirements.txt
altair==5.0.1
attrs==24.2.0
backports.zoneinfo==0.2.1
blinker==1.6.3
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
importlib-metadata==6.7.0
importlib-resources==5.12.0
jinja2==3.1.4
jsonschema==4.17.3
markdown-it-py==2.2.0
MarkupSafe==2.1.5
mdurl
numpy
packaging==23.2
pandas==1.1.5
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
protobuf==4.24.4
pyarrow==12.0.1
pydeck==0.8.1b1
pygments==2.17.2
Pympler==1.1
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
pywin32==306
requests==2.31.0
rich==13.7.1
six==1.16.0
smmap==5.0.1
streamlit==1.23.1
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.2
typing-extensions==4.7.1
tzdata==2024.1
tzlocal==4.3.1
urllib3==2.0.7
validators==0.20.0
watchdog==3.0.0
zipp==3.15.0
